Setting up Redux.
client/store.js : creates the store from which all components will access state.

App.js : provider wrapped around the router and instantiated with the store created to provide state to the components in the Router.

client/constants/actionTypes: adding to specify which actions are allowed.

client/actions/auth.js : loadUser and login functions are created to interact with the login component. login will eventually make a request to the backend for the token. Once the token is received, loadUser will load the user's data (again, in the future, a GET request will be sent to the /user route for example with the header 'x-auth-token'). The login function dispatches the appropriate action (including the user data as a payload) to the reducer depending on the result.

authReducer handles the USER_LOADED, REGISTER_SUCCESS, LOGIN_SUCCESS, and LOGOUT actions for now. There is code that will eventually handle registration as well, but they are similar to LOGIN_SUCCESS and LOGOUT.

client/reducers/index.js combines all the reducers.

client/components/Login.js is connected to the provider to get the pieces of state that it needs and the login function is mapped to the properties and is a good example to understand how to connect both state and functions to the properties.

client/data/data.js contains an object that has all of the sample data