Adding some front end styling
Landing is using a high quality photo in the background and the navbar is a cleaner flexbox.

The Course component was created to organize the dashboard and will be dynamically rendered based on the user's data in the backend.

Login and Sign Up forms' CSS was cleaned up to look more presentable.

Assets were added to the images folder to be used as backgrounds for components.